refactor: merge page iteration with codegen
