+++ projects/angular-codegen/src/public_api.ts
@@ -5,6 +5,5 @@
 export * from './lib/element/angular-element-codegen.service';
 export * from './lib/element/angular-element-codegen.module';
 export * from './lib/angular-docgen.service';
-export * from './lib/angular-bootstrap.service';
 export * from './lib/angular-codegen.service';
 export * from './lib/angular-codegen.module';
