+++ projects/stencil-codegen/src/lib/stencil-codegen.service.ts
@@ -25,12 +25,10 @@ export class StencilCodeGenService {
     this.webCodeGen.compute(current, data, this.compileOptions(options));
   }
 
-  aggregate(
-    current: SketchMSLayer,
-    data: SketchMSData,
-    options?: WebCodeGenOptions
-  ) {
-    return this.visit(current, data, this.compileOptions(options));
+  aggregate(data: SketchMSData, options?: WebCodeGenOptions) {
+    return data.pages.flatMap(page =>
+      this.visit(page, data, this.compileOptions(options))
+    );
   }
 
   identify(current: SketchMSLayer) {
