+++ src/app/editor/code/editor-container/codegen/angular-codegen.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import {
-  AngularBootstrapService,
   AngularCodeGenService,
   AngularDocGenService
 } from '@xlayers/angular-codegen';
@@ -11,7 +10,6 @@ import { XlayersNgxEditorModel } from './codegen.service';
 })
 export class AngularCodeGenFacadeService {
   constructor(
-    private readonly angularBootstrapService: AngularBootstrapService,
     private readonly angularCodeGen: AngularCodeGenService,
     private readonly angularDocGen: AngularDocGenService
   ) {}
@@ -23,14 +21,8 @@ export class AngularCodeGenFacadeService {
   }
 
   generate(data: SketchMSData) {
-    const files = data.pages.flatMap(page =>
-      this.angularCodeGen.aggregate(page, data)
-    );
-
-    return [
-      ...this.angularDocGen.aggregate(data),
-      ...this.angularBootstrapService.generate(files),
-      ...files
-    ] as XlayersNgxEditorModel[];
+    return this.angularDocGen
+      .aggregate(data)
+      .concat(this.angularCodeGen.aggregate(data)) as XlayersNgxEditorModel[];
   }
 }
